+++ test/expression/parse.test.js
@@ -981,8 +981,20 @@ describe('parse', function() {
         assert.equal(parseAndEval('+4!'), 24);
       });
 
-      it.skip('should respect precedence of transpose', function () {
-        // TODO: test transpose
+      it('should respect precedence of transpose', function () {
+        var node = math.parse('a + b\'');
+        assert(node instanceof OperatorNode);
+        assert.equal(node.op, '+');
+        assert.equal(node.params[0].toString(), 'a');
+        assert.equal(node.params[1].toString(), 'b\'');
+      });
+
+      it('should respect precedence of transpose (2)', function () {
+        var node = math.parse('a ^ b\'');
+        assert(node instanceof OperatorNode);
+        assert.equal(node.op, '^');
+        assert.equal(node.params[0].toString(), 'a');
+        assert.equal(node.params[1].toString(), 'b\'');
       });
 
       it('should respect precedence of conditional operator and other operators', function () {
@@ -993,14 +1005,6 @@ describe('parse', function() {
         assert.deepEqual(parseAndEval('3 ? 5cm to m : 5cm in mm'), new Unit(5, 'cm').to('m'));
         assert.deepEqual(parseAndEval('2 == 4-2 ? [1,2] : false'), new Matrix([1,2]));
         assert.deepEqual(parseAndEval('false ? 1:2:6'), new Matrix([2,3,4,5,6]));
-      });
-
-      it('should respect precedence of equal operator and conversion operators', function () {
-        var node = math.parse('a == b to c'); // (a == b) to c
-        assert.equal(node.op, 'to');
-
-        var node2 = math.parse('a to b == c');
-        assert.equal(node2.op, 'to');
       });
 
       it('should respect precedence of conditional operator and relational operators', function () {
@@ -1047,6 +1051,20 @@ describe('parse', function() {
         assert(node instanceof RangeNode);
         assert.equal(node.start.toString(), 'a + b');
         assert.equal(node.end.toString(), 'c - d');
+      });
+
+      it('should respect precedence of "to" operator and relational operators', function () {
+        var node = math.parse('a == b to c');
+        assert(node instanceof OperatorNode);
+        assert.equal(node.params[0].toString(), 'a');
+        assert.equal(node.params[1].toString(), 'b to c');
+      });
+
+      it('should respect precedence of "to" operator and relational operators (2)', function () {
+        var node = math.parse('a to b == c');
+        assert(node instanceof OperatorNode);
+        assert.equal(node.params[0].toString(), 'a to b');
+        assert.equal(node.params[1].toString(), 'c');
       });
 
       // TODO: extensively test operator precedence